PERF move to_camelcase to the only module it's used
